Moved reportInternalError and reportError to express-error handler
